move HttpSSLConfigServer to fetcher
